Tokenizer: Add support for decimal and exponential numbers

+++ src/parser/value.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 #[derive(Debug)]
 pub enum JsonValue {
     String(String),
-    Number(i64),
+    Number(f64),
     Boolean(bool),
     Null,
